KeyAlgorithm moved to mod jwk, typo fixes
