+++ src/lib.rs
@@ -17,7 +17,7 @@ mod pem;
 mod serialization;
 mod validation;
 
-pub use algorithms::{Algorithm, KeyAlgorithm};
+pub use algorithms::Algorithm;
 pub use decoding::{decode, decode_header, DecodingKey, TokenData};
 pub use encoding::{encode, EncodingKey};
 pub use header::Header;
